Lint fixes (#1688)
* Update golangci-lint config

Use the same linters as opentelemetry-go

* Apply auto-fixes

* Fix errorlint issues

* Fix unconvert issues

Remove unnecessary conversions

* Fix gosec lint errors

* Ignore lint error in httpPlusdb example

* Remove unnecessary conv in kafka/consumer example

* Ignore lint warning in kafka-go/producer example

* Fix lint in examples/rolldice/frontend

* Fix lint in examples/rolldice/user/server

* Ignore lint error in databasesql e2e test

* Ignore lint error in kafka-go e2e test

* Ignore lint error in nethttp e2e test

* Fix lint in nethttp_custom e2e test

* Fix lint in internal/inspect

* Fix gosec issues in sdk

* Fixes

* Fix remaining testify issues

* Fix issues with span tests in SDK

- Run the function returned from assertTracer.
- Call the span passed to the goroutine.

* No check for raw decode of attr

* Remove now unneeded nolint

* Remove unnecessary conv and use MaxUint64

+++ examples/kafka-go/consumer/main.go
@@ -48,7 +48,7 @@ func reader(ctx context.Context) {
 			span.SetAttributes(
 				attribute.String("topic", m.Topic),
 				attribute.Int64("partition", int64(m.Partition)),
-				attribute.Int64("offset", int64(m.Offset)),
+				attribute.Int64("offset", m.Offset),
 			)
 			fmt.Printf("consumed message at topic:%v partition:%v offset:%v	%s = %s\n", m.Topic, m.Partition, m.Offset, string(m.Key), string(m.Value))
 			span.End()
